+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/_validators.py
@@ -9,7 +9,7 @@ from knack.util import CLIError
 
 def validate_volume_mount_path(ns):
     if ns.azure_file_volume_mount_path and ':' in ns.azure_file_volume_mount_path:
-        raise CLIError("The Azure File volume mount path cannot contain ':'")
+        raise CLIError("The volume mount path cannot contain ':'")
 
 
 def validate_secrets(ns):
@@ -30,3 +30,8 @@ def validate_secret(string):
             raise CLIError("Secrets need to be specifed in key=value format.")
         result = {comps[0]: b64encode(comps[1].encode('ascii')).decode('ascii')}
     return result
+
+
+def validate_gitrepo_directory(ns):
+    if ns.gitrepo_dir and '..' in ns.gitrepo_dir:
+        raise CLIError("The git repo directory cannot contain '..'")